+++ src/scala/de/htwg/winesmeeper/Main.scala
@@ -1,9 +1,0 @@
-package de.htwg.winesmeeper
-
-import de.htwg.winesmeeper.Config
-
-val initVals = new Array[Int](5)
-  
-@main def start: Unit =
-  val gb = Config.startBoard
-  val ctrl = Config.startController(gb)